fix: double reducer bug fixed
