+++ src/store/reducers/reducers.js
@@ -1,8 +1,9 @@
 //imports
 import { combineReducers } from 'redux';
-import operations from './operations'
+import operations from './operations';
 //initial states
 let userLoggedIn = {};
+let userWallets = {};
 
 //reducers
 const userLogged = (state = userLoggedIn, action) => {
@@ -18,8 +19,20 @@ const userLogged = (state = userLoggedIn, action) => {
   }
 };
 
+const getWallets = (state = userWallets, action) => {
+  switch (action.type) {
+  case 'GET_WALLETS':
+    console.log(action.data, 'XXXX')
+    userWallets = action.data;
+    return userWallets;
+  default:
+    return state;
+
+  }
+};
 //export reducers
 export const reducers = combineReducers({
   userLogged,
-  operations
+  operations,
+  getWallets
 });
